+++ tests/Commands/__snapshots__/ModuleDeleteCommandTest__it_deletes_modules_from_status_file__1.php
@@ -1,0 +1,3 @@
+<?php return '{
+    "WrongModule": true
+}';